subscriber/src/requirements.txt
asn1tools==0.166.0
deprecation==2.1.0
expiring_dict==1.1.0
Flask==3.0.1
flask_mqtt==1.1.1
flask_pymongo==2.3.0
geojson==3.1.0
haversine==2.8.1
jsonschema==4.21.1
ntplib==0.4.0
python_logstash_async==2.7.2
paho_mqtt==1.6.1
pyngsi==2.1.10
setuptools==59.6.0
diskcache==5.6.3

matplotlib==3.9.2
matplotlib-inline==0.1.7
numpy==1.26.4
open3d==0.18.0
opencv-python==4.10.0.84
#PyAudio==0.2.14
#pyinput==0.3.2
pynput==1.7.6
av==13.1.0
jproperties==2.1.0
argparse==1.4.0
pyzmq==26.2.0
eclipse-zenoh==1.2.1
json5==0.10.0